+++ db/mem_batch.go
@@ -1,6 +1,8 @@
 package db
 
-import "sync"
+import (
+	"sync"
+)
 
 type atomicSetDeleter interface {
 	Mutex() *sync.Mutex
@@ -66,6 +68,5 @@ func (mBatch *memBatch) write(doSync bool) {
 		case opTypeDelete:
 			mBatch.db.DeleteNoLock(op.key)
 		}
-
 	}
 }