+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 0.8.2 (April 12th, 2018)
+
+FEATURES:
+
+ - [db] DebugDB shows better colorized output
+
+BUG FIXES:
+
+ - [db] PrefixDB Iterator/ReverseIterator fixes
+ - [db] DebugDB fixes
+
 ## 0.8.1 (April 5th, 2018)
 
 FEATURES: